Small code refactoring Increased HTTPS security
